Use distributed scheduler
