+++ Crystalball/budget.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 import psutil
 import numpy as np
 
@@ -5,10 +7,7 @@ import numpy as np
 def get_budget(nr_sources, nr_rows, nr_chans, nr_corrs, data_type, cb_args,
                fudge_factor=2., row2source_ratio=100):
     systmem = np.float(psutil.virtual_memory()[0])
-    if not cb_args.num_workers:
-        nrthreads = psutil.cpu_count()
-    else:
-        nrthreads = cb_args.num_workers
+    nrthreads = cb_args.num_workers
 
     print('-------------------------------------------')
     print('system RAM = {0:.2f} GB'.format(systmem/1024**3))
